Modified details fragment

+++ app/src/main/java/com/example/cryptowatch/models/CoinHistory.kt
@@ -1,7 +1,7 @@
 package com.example.cryptowatch.models
 
 data class CoinHistory(
-    val market_caps: List<List<Long>>,
-    val prices: List<List<Long>>,
-    val total_volumes: List<List<Long>>
+    val market_caps: List<List<Double>>,
+    val prices: List<List<Double>>,
+    val total_volumes: List<List<Double>>
 )
